+++ test/HCurve.test.js
@@ -26,7 +26,6 @@ const {
   CURVE_Y_DEPOSIT,
   CURVE_SBTC_SWAP,
   CURVE_SETH_SWAP,
-  CURVE_ONE_SPLIT,
   CURVE_YCRV,
   CURVE_YCRV_PROVIDER,
   CURVE_SBTCCRV,
@@ -48,7 +47,6 @@ const Proxy = artifacts.require('ProxyMock');
 const Registry = artifacts.require('Registry');
 const HCurve = artifacts.require('HCurve');
 const ICurveHandler = artifacts.require('ICurveHandler');
-const IOneSplit = artifacts.require('IOneSplit');
 const IToken = artifacts.require('IERC20');
 const IYToken = artifacts.require('IYToken');
 
@@ -68,7 +66,6 @@ contract('Curve', function([_, user]) {
     this.sbtcSwap = await ICurveHandler.at(CURVE_SBTC_SWAP);
     this.sethSwap = await ICurveHandler.at(CURVE_SETH_SWAP);
     this.aaveSwap = await ICurveHandler.at(CURVE_AAVE_SWAP);
-    this.oneSplit = await IOneSplit.at(CURVE_ONE_SPLIT);
   });
 
   beforeEach(async function() {
@@ -434,149 +431,6 @@ contract('Curve', function([_, user]) {
         );
         expect(await this.token.balanceOf.call(user)).to.be.bignumber.eq(
           tokenUser
-        );
-        profileGas(receipt);
-      });
-    });
-  });
-
-  describe('OneSplit swap', function() {
-    const token0Address = SUSD_TOKEN;
-    const token1Address = TUSD_TOKEN;
-    const providerAddress = SUSD_PROVIDER;
-
-    let token0User;
-    let token1User;
-
-    before(async function() {
-      this.token0 = await IToken.at(token0Address);
-      this.token1 = await IToken.at(token1Address);
-    });
-
-    beforeEach(async function() {
-      token0User = await this.token0.balanceOf.call(user);
-      token1User = await this.token1.balanceOf.call(user);
-    });
-
-    describe('susd to y pool through oneSplit', function() {
-      it('Exact input swap sUSD to TUSD by OneSplit', async function() {
-        const value = ether('1');
-        const parts = new BN('2');
-        const flags = 0x401e006d000;
-        const answer = await this.oneSplit.getExpectedReturn.call(
-          this.token0.address,
-          this.token1.address,
-          value,
-          parts,
-          flags,
-          {
-            from: user,
-          }
-        );
-        const data = abi.simpleEncode(
-          'swap(address,address,uint256,uint256,uint256[],uint256)',
-          this.token0.address,
-          this.token1.address,
-          value,
-          mulPercent(answer.returnAmount, new BN('100').sub(slippage)),
-          answer.distribution,
-          flags
-        );
-        await this.token0.transfer(this.proxy.address, value, {
-          from: providerAddress,
-        });
-        await this.proxy.updateTokenMock(this.token0.address);
-        const receipt = await this.proxy.execMock(this.hCurve.address, data, {
-          from: user,
-          value: ether('1'), // Ensure handler can correctly deal with ether
-        });
-
-        // Get handler return result
-        const handlerReturn = utils.toBN(
-          getHandlerReturn(receipt, ['uint256'])[0]
-        );
-        const token1UserEnd = await this.token1.balanceOf.call(user);
-        expect(handlerReturn).to.be.bignumber.eq(token1UserEnd.sub(token1User));
-
-        expect(
-          await this.token0.balanceOf.call(this.proxy.address)
-        ).to.be.bignumber.eq(ether('0'));
-        expect(
-          await this.token1.balanceOf.call(this.proxy.address)
-        ).to.be.bignumber.eq(ether('0'));
-        expect(await this.token0.balanceOf.call(user)).to.be.bignumber.eq(
-          token0User
-        );
-        // oneSplit use sUSD and y pools in this case, give 10% tolerance
-        // for sUSD/TUSD.
-        expect(await this.token1.balanceOf.call(user)).to.be.bignumber.gte(
-          token1User
-            .add(answer.returnAmount)
-            .sub(answer.returnAmount.div(new BN('10')))
-        );
-        expect(await this.token1.balanceOf.call(user)).to.be.bignumber.lte(
-          token1User.add(answer.returnAmount)
-        );
-        profileGas(receipt);
-      });
-
-      it('Exact input swap sUSD to TUSD by OneSplit with max amount', async function() {
-        const value = ether('1');
-        const parts = new BN('2');
-        const flags = 0x401e006d000;
-        const answer = await this.oneSplit.getExpectedReturn.call(
-          this.token0.address,
-          this.token1.address,
-          value,
-          parts,
-          flags,
-          {
-            from: user,
-          }
-        );
-        const data = abi.simpleEncode(
-          'swap(address,address,uint256,uint256,uint256[],uint256)',
-          this.token0.address,
-          this.token1.address,
-          MAX_UINT256,
-          mulPercent(answer.returnAmount, new BN('100').sub(slippage)),
-          answer.distribution,
-          flags
-        );
-        await this.token0.transfer(this.proxy.address, value, {
-          from: providerAddress,
-        });
-        await this.proxy.updateTokenMock(this.token0.address);
-        const receipt = await this.proxy.execMock(this.hCurve.address, data, {
-          from: user,
-          value: ether('1'), // Ensure handler can correctly deal with ether
-        });
-
-        // Get handler return result
-        const handlerReturn = utils.toBN(
-          getHandlerReturn(receipt, ['uint256'])[0]
-        );
-        const token1UserEnd = await this.token1.balanceOf.call(user);
-        expect(handlerReturn).to.be.bignumber.eq(token1UserEnd.sub(token1User));
-
-        expect(
-          await this.token0.balanceOf.call(this.proxy.address)
-        ).to.be.bignumber.eq(ether('0'));
-        expect(
-          await this.token1.balanceOf.call(this.proxy.address)
-        ).to.be.bignumber.eq(ether('0'));
-        expect(await this.token0.balanceOf.call(user)).to.be.bignumber.eq(
-          token0User
-        );
-        // oneSplit use sUSD and y pools in this case, give 10% tolerance
-        // for sUSD/TUSD.
-        expect(await this.token1.balanceOf.call(user)).to.be.bignumber.gte(
-          token1User
-            .add(answer.returnAmount)
-            .sub(answer.returnAmount.div(new BN('10')))
-        );
-        expect(await this.token1.balanceOf.call(user)).to.be.bignumber.lte(
-          token1User.add(answer.returnAmount)
         );
         profileGas(receipt);
       });